fix: update MediaPreview and add MediaPreviewContent
