+++ utils/fileUtils.ts
@@ -1,0 +1,6 @@
+export const formatFileSize = (bytes?: number | null): string => {
+  if (!bytes) return '';
+  const sizes = ['B', 'KB', 'MB', 'GB'];
+  const i = Math.floor(Math.log(bytes) / Math.log(1024));
+  return `${(bytes / Math.pow(1024, i)).toFixed(1)} ${sizes[i]}`;
+};
